Move less used methods and experiments out of main Pipelines project
