+++ corefxlab.sln
@@ -1,6 +1,6 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26226.0
+VisualStudioVersion = 15.0.26228.4
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{5E7EB061-B9BC-4DA2-B5E5-859AA7C67695}"
 	ProjectSection(SolutionItems) = preProject
@@ -131,6 +131,8 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "System.Text.Primitives.Performance.Tests", "tests\System.Text.Primitives.Performance.Tests\System.Text.Primitives.Performance.Tests.csproj", "{04970622-6C7A-4FB4-814A-9BFE051A2E68}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "System.IO.Pipelines.Performance.Tests", "tests\System.IO.Pipelines.Performance.Tests\System.IO.Pipelines.Performance.Tests.csproj", "{E68CF803-CC9F-458A-B0C7-1C2E0A0A3E4A}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "System.IO.Pipelines.Extensions", "src\System.IO.Pipelines.Extensions\System.IO.Pipelines.Extensions.csproj", "{518A169E-8EA6-40CA-8729-B9BCF6FC4CD7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -850,6 +852,18 @@ Global
 		{E68CF803-CC9F-458A-B0C7-1C2E0A0A3E4A}.Release|x64.Build.0 = Release|Any CPU
 		{E68CF803-CC9F-458A-B0C7-1C2E0A0A3E4A}.Release|x86.ActiveCfg = Release|Any CPU
 		{E68CF803-CC9F-458A-B0C7-1C2E0A0A3E4A}.Release|x86.Build.0 = Release|Any CPU
+		{518A169E-8EA6-40CA-8729-B9BCF6FC4CD7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{518A169E-8EA6-40CA-8729-B9BCF6FC4CD7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{518A169E-8EA6-40CA-8729-B9BCF6FC4CD7}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{518A169E-8EA6-40CA-8729-B9BCF6FC4CD7}.Debug|x64.Build.0 = Debug|Any CPU
+		{518A169E-8EA6-40CA-8729-B9BCF6FC4CD7}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{518A169E-8EA6-40CA-8729-B9BCF6FC4CD7}.Debug|x86.Build.0 = Debug|Any CPU
+		{518A169E-8EA6-40CA-8729-B9BCF6FC4CD7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{518A169E-8EA6-40CA-8729-B9BCF6FC4CD7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{518A169E-8EA6-40CA-8729-B9BCF6FC4CD7}.Release|x64.ActiveCfg = Release|Any CPU
+		{518A169E-8EA6-40CA-8729-B9BCF6FC4CD7}.Release|x64.Build.0 = Release|Any CPU
+		{518A169E-8EA6-40CA-8729-B9BCF6FC4CD7}.Release|x86.ActiveCfg = Release|Any CPU
+		{518A169E-8EA6-40CA-8729-B9BCF6FC4CD7}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -914,5 +928,6 @@ Global
 		{1ED92192-6C2A-44A5-9098-F9CBABB0292D} = {3079E458-D0E6-4F99-8CAB-80011D35C7DA}
 		{04970622-6C7A-4FB4-814A-9BFE051A2E68} = {3079E458-D0E6-4F99-8CAB-80011D35C7DA}
 		{E68CF803-CC9F-458A-B0C7-1C2E0A0A3E4A} = {3079E458-D0E6-4F99-8CAB-80011D35C7DA}
+		{518A169E-8EA6-40CA-8729-B9BCF6FC4CD7} = {4B000021-5278-4F2A-B734-DE49F55D4024}
 	EndGlobalSection
 EndGlobal
