+++ samples/System.IO.Pipelines.Samples/HttpServer/HttpServer.cs
@@ -176,4 +176,6 @@ namespace System.IO.Pipelines.Samples.Http
             await connection.ProcessAllRequests();
         }
     }
+
+
 }
