+++ src/System.IO.Pipelines.Extensions/PipelineConnectionExtensions.cs
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace System.IO.Pipelines
+{
+    public static class PipelineConnectionExtensions
+    {
+        public static Stream GetStream(this IPipeConnection connection)
+        {
+            return new PipelineConnectionStream(connection);
+        }
+    }
+}
