+++ src/System.IO.Pipelines.Extensions/PipelineConnectionStream.cs
@@ -26,37 +26,22 @@ namespace System.IO.Pipelines
 
         public override bool CanWrite => true;
 
-        public override long Length
-        {
-            get
-            {
-                ThrowHelper.ThrowNotSupportedException();
-                return 0;
-            }
-        }
+        public override long Length => throw new NotSupportedException();
 
         public override long Position
         {
-            get
-            {
-                ThrowHelper.ThrowNotSupportedException();
-                return 0;
-            }
-            set
-            {
-                ThrowHelper.ThrowNotSupportedException();
-            }
+            get => throw new NotSupportedException();
+            set => throw new NotSupportedException();
         }
 
         public override long Seek(long offset, SeekOrigin origin)
         {
-            ThrowHelper.ThrowNotSupportedException();
-            return 0;
+            throw new NotSupportedException();
         }
 
         public override void SetLength(long value)
         {
-            ThrowHelper.ThrowNotSupportedException();
+            throw new NotSupportedException();
         }
         public override int Read(byte[] buffer, int offset, int count)
         {
