+++ src/System.IO.Pipelines.Extensions/PipelineReaderExtensions.cs
@@ -1,0 +1,40 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace System.IO.Pipelines
+{
+    public static class PipelineReaderExtensions
+    {
+
+        public static Task CopyToAsync(this IPipeReader input, Stream stream)
+        {
+            return input.CopyToAsync(stream, 4096, CancellationToken.None);
+        }
+
+        public static async Task CopyToAsync(this IPipeReader input, Stream stream, int bufferSize, CancellationToken cancellationToken)
+        {
+            // TODO: Use bufferSize argument
+            while (!cancellationToken.IsCancellationRequested)
+            {
+                var result = await input.ReadAsync();
+                var inputBuffer = result.Buffer;
+                try
+                {
+                    if (inputBuffer.IsEmpty && result.IsCompleted)
+                    {
+                        return;
+                    }
+
+                    await inputBuffer.CopyToAsync(stream);
+                }
+                finally
+                {
+                    input.Advance(inputBuffer.End);
+                }
+            }
+        }
+    }
+}
