+++ src/System.IO.Pipelines.Extensions/StreamPipeConnection.cs
@@ -1,9 +1,6 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
-using System.IO;
-
 namespace System.IO.Pipelines
 {
     internal class StreamPipeConnection : IPipeConnection
