+++ src/System.IO.Pipelines.Extensions/System.IO.Pipelines.Extensions.csproj
@@ -1,0 +1,11 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <Import Project="..\..\tools\common.props" />
+  <PropertyGroup>
+    <Description>Stream pipeline adapters implementation</Description>
+    <TargetFrameworks>net451;netstandard1.3</TargetFrameworks>
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+  </PropertyGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\System.IO.Pipelines\System.IO.Pipelines.csproj" />
+  </ItemGroup>
+</Project>
