+++ src/System.IO.Pipelines/IPipeReader.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
-
 namespace System.IO.Pipelines
 {
     /// <summary>
