+++ src/System.IO.Pipelines/IPipeWriter.cs
@@ -1,9 +1,6 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
-using System.Threading.Tasks;
-
 namespace System.IO.Pipelines
 {
     /// <summary>
