+++ src/System.IO.Pipelines/IReadableBufferAwaiter.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
-
 namespace System.IO.Pipelines
 {
     public interface IReadableBufferAwaiter
@@ -10,15 +8,6 @@ namespace System.IO.Pipelines
         bool IsCompleted { get; }
 
         ReadResult GetResult();
-
-        void OnCompleted(Action continuation);
-    }
-
-    public interface IWritableBufferAwaiter
-    {
-        bool IsCompleted { get; }
-
-        bool GetResult();
 
         void OnCompleted(Action continuation);
     }
