+++ src/System.IO.Pipelines/InlineScheduler.cs
@@ -2,12 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
 namespace System.IO.Pipelines
 {
     public class InlineScheduler : IScheduler
