+++ src/System.IO.Pipelines/WritableBuffer.cs
@@ -1,9 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
 using System.Buffers;
-using System.Threading.Tasks;
 
 namespace System.IO.Pipelines
 {
